Simplifying Events
Currently, almost anything that can happen to an LP in the model is
passes either as an event from itself to itself or from another lp to
it. This branch is an attempt to restructure internal events such as
expanding, scaling (up and down), and rebuilding such that they are not
actually passed as ROSS events anymore. This way, timestamp ties might
be reduced.