Safety Check
Added a safeguard to set the offense value to 0 in case an LP tries to
downscale too much. Also made sure that the LP can’t get any free
resources by hitting (and passing) the bottom. Tests now run to
completion without crashes from this issue. The number of conflicting
timestamps in a 4 core, 50 LPs-per-core test was 35,966, rather than
between 600,000 to 1 million!