Print Output
When the stats for each LP were printed, the printout would often print
over itself and in general be unreadable in places. This commit’s main
purpose is to add a mechanism to fix this. Implementation: a global
array of structs, each of which contain the data for each LP. This
global array is modified by the statistics collection function, which
ROSS calls for every LP, and then main will sequentially print
everything out on its own.