Reverse Handler Works
This commit doesn’t really fix the handler, since it was already
working. Outputs were never being recorded by ROSS in optimistic debug
mode, which is why it never appeared to be working. Will keep the misc
changes in the commit, thought the ROSS source code change that allowed
the output to be recorded and outputted (showing that the reverse
handler works) was the addition of the call ‘tw_stats(me);’ on line 498
of tw-sched.c